requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Claude API (judgement analysis only - not used for scraping/parsing)
anthropic>=0.40.0

# Telegram
python-telegram-bot>=20.7

# Scheduling
schedule>=1.2.0

# Logging
python-json-logger>=2.0.0
